refactore server
