+++ package.json
@@ -4,15 +4,20 @@
   "description": "An app that allows events centers managers to manage users applications to use their centers",
   "main": "app.js",
   "scripts": {
-    "migrate:dev": "sequelize db:migrate",
-    "unmigrate:dev": "sequelize db:migrate:undo:all",
-    "migrate:test": "sequelize db:migrate --env test",
-    "unmigrate:test": "sequelize db:migrate:undo:all --env test",
-    "seed": "sequelize db:seed:all",
-    "unseed": "sequelize db:seed:undo:all",
-    "start:dev": "nodemon --watch server --exec babel-node ./bin/www",
-    "start:prod": "babel-node ./bin/www",
-    "test": "SET NODE_ENV=test && mocha --compilers js:babel-core/register server/test/**.test/*.test.js --timeout 20000"
+
+    "migrate:dev": "node ./node_modules/sequelize-cli/lib/sequelize db:migrate",
+    "heroku-postbuild": "npm run migrate:dev && npm run build-server && npm run seed",
+    "unmigrate:dev": "node ./node_modules/sequelize-cli/lib/sequelize db:migrate:undo:all",
+    "migrate:test": "node ./node_modules/sequelize-cli/lib/sequelize db:migrate --env test",
+    "unmigrate:test": "node ./node_modules/sequelize-cli/lib/sequelize db:migrate:undo:all --env test",
+    "seed": "node ./node_modules/sequelize-cli/lib/sequelize db:seed:all",
+    "unseed": "node ./node_modules/sequelize-cli/lib/sequelize db:seed:undo:all",
+    "start:dev": "nodemon --watch server --exec babel-node ./server/bin/www",
+    "start": "node ./build/bin/www.js",
+    "build-server": " node ./node_modules/babel-cli/bin/babel.js -d ./build ./server",
+    "test": "nyc mocha --compilers js:babel-core/register server/test/**.test/*.test.js --timeout 60000 --exit",
+    "tests": "SET NODE_ENV=test && mocha --compilers js:babel-core/register server/test/**.test/*.test.js --timeout 60000 --exit && npm run coverage",
+    "coverage": "npm run test && nyc report --reporter=text-lcov"
   },
   "author": "Kingsley Frank-Demesi",
   "license": "ISC",
@@ -44,13 +49,17 @@
     "pg": "^7.4.0",
     "pg-hstore": "^2.3.2",
     "sequelize": "^4.22.8",
+    "sequelize-cli": "^3.1.0",
     "supertest": "^3.0.0"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-watch": "^2.0.7",
+    "coveralls": "^3.0.0",
     "keypair": "^1.0.1",
-    "nodemon": "^1.12.1"
+    "mocha-lcov-reporter": "^1.3.0",
+    "nodemon": "^1.12.1",
+    "nyc": "^11.3.0"
   },
   "repository": {
     "type": "git",
