+++ server/test/services.test/validationService.test.js
@@ -1,7 +1,7 @@
 import chai from 'chai';
 import supertest from 'supertest';
 import ValidationService from '../../services/ValidationService';
-import app from '../../../app.js';
+import app from '../../app.js';
 
 const expect = chai.expect,
 			server = supertest.agent(app);
